get started with a devcontainer
this setup follows this blog post [1], but the blog post has a bug in it
where it builds a virtualenv in the workspace directory that eventually
gets blown up when the container starts up. The blog post tries to
create a dedicated virtualenv instead of the py-utils environment (which
is sort of like a "global" environment for the container) which seems
like a reasonable way to go about build an isolated dev environment, but
I'm sort of new to devcontainers, but it doesn't look like this is the
standard way to do things.

I'll probably poke around some documentation some more, but this seems
like a good place to start

[1]: https://stemmetje.com/2022/11/dev-container-setup-for-python/