build virtualenv in workspace dir
sometimes this makes poetry/pip complain that the virtualenv is broken
this is probably because the .venv/pyvenv.cfg file doesn't show the
right environment info (e.g. it shows the devcontainer file structure
or the file structure of a physical machine). not a huge deal, since
poetry recreates the cfg file with the appropriate config data