close unused tasks
we create a wait-for-shutdown-condition task every iteration of the
loop. we should cancel these tasks when we haven't received a shutdown
condition notify.

we also create a new smartbridge topology, and if we leave these
topology connections around, we'll saturate the connections to the
smartbridge and it will stop behaving